pyproject.toml
[project]
name = "oh-my-port"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyqt6>=6.11.0",
    "pyserial>=3.5",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]
